chore: rename config_util to config
